tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/cssmediarule-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSMediaRule - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssmediarule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSMediaRule' in window);

      }, '\'CSSMediaRule\' in window');

      test(function(){

        assert_true(CSSMediaRule instanceof Function);

      }, 'CSSMediaRule instanceof Function');

      test(function(){

        assert_true('prototype' in CSSMediaRule);

      }, '\'prototype\' in CSSMediaRule');

      test(function(){

        assert_equals(CSSMediaRule.prototype.toString(), '[object CSSMediaRulePrototype]');

      }, 'CSSMediaRule.prototype.toString() == \'[object CSSMediaRulePrototype]\'');

      test(function(){

        assert_true('constructor' in CSSMediaRule.prototype);

      }, '\'constructor\' in CSSMediaRule.prototype');

      test(function(){

        assert_equals(CSSMediaRule.prototype.constructor,CSSMediaRule);

      }, 'CSSMediaRule.prototype.constructor === CSSMediaRule');

    ]]>
    </script>
  </body>
</html>
